Fix cache tail eviction and add tests
