+++ archiver/mrucache_test.go
@@ -8,18 +8,19 @@ import (
 func TestMRUCache(t *testing.T) {
 	var cache *Cache = NewCache(8)
 
-	// Fill cache with values 0,1,2,3,4,5,6,9
 	for i := 0; i < 10; i++ {
 		cache.Set(fmt.Sprintf("testkey%d", i), i)
+		cache.Get(fmt.Sprintf("testkey%d", i))
 	}
+	// Cache should have values 2,3,4,5,6,7,8,9
 	fmt.Println(cache.cache)
 
 	h := cache.head
 	for {
+		fmt.Printf("at element k %v v %v\n", h.key, h.value)
 		if h == cache.tail {
 			break
 		}
-		fmt.Printf("at element k %v v %v\n", h.key, h.value)
 		h = h.prev
 	}
 
@@ -27,7 +28,7 @@ func TestMRUCache(t *testing.T) {
 	for i := 2; i < 10; i++ {
 		val, success := cache.Get(fmt.Sprintf("testkey%d", i))
 		if val != i || !success {
-			t.Errorf("Missing value %d in cache", i)
+			t.Errorf("value %d should be in cache, but is not", i)
 		}
 	}
 
@@ -130,5 +131,102 @@ func TestMRUCacheTestFirstEviction(t *testing.T) {
 		}
 		current = current.next
 		index++
+	}
+}
+
+func TestGet(t *testing.T) {
+	lru := NewCache(uint32(4))
+
+	val, ok := lru.Get("asdf")
+	if ok != false {
+		t.Error("ok should be false but is", ok)
+	}
+	lru.Set("asdf", "asdfvalue")
+	val, ok = lru.Get("asdf")
+	if val != "asdfvalue" {
+		t.Error("LRU.Get does not return correct value", val)
+	}
+
+	if lru.cache["asdf"].value.(string) != "asdfvalue" {
+		t.Error("LRU.cache does not contain key/value after Get")
+	}
+
+}
+
+func TestEviction(t *testing.T) {
+	lru := NewCache(2)
+	val1, ok := lru.Get("a")
+	if ok != false {
+		t.Error("ok should be false")
+	}
+	lru.Set("a", "avalue")
+	val1, ok = lru.Get("a")
+	if ok != true {
+		t.Error("ok should be true")
+	}
+	if val1 != "avalue" {
+		t.Error("lru.Get: val1 should be avalue but is", val1)
+	}
+
+	lru.Set("b", "bvalue")
+	lru.Set("c", "cvalue")
+
+	if len(lru.cache) != 2 {
+		t.Error("lru.Cache size should be 2 but is", len(lru.cache))
+	}
+
+	val1, ok = lru.Get("a")
+	if ok == true {
+		t.Error("a should have been evicted")
+	}
+
+	val1, ok = lru.Get("b")
+	if ok != true {
+		t.Error("b should have been retained")
+	}
+
+	val1, ok = lru.Get("c")
+	if ok != true {
+		t.Error("c should have been retained")
+	}
+}
+
+/**
+ * Should do the following benchmarks:
+ * Insert with no Reuse (no repeats)
+ * Insert with Resue (with repeats -- moving LRU item to top)
+ * Parallel versions of the above, to test the effect of the Lock
+ * Insert with LRU size = 1
+ * Insert+Get with LRU size = 1
+ * Get with LRU size = 1
+**/
+
+func BenchmarkSetSize1NoReuse(b *testing.B) {
+	l := NewCache(uint32(3))
+	for i := 0; i < b.N; i++ {
+		l.Set(string(i), i)
+	}
+}
+
+func BenchmarkSetSize1000Reuse(b *testing.B) {
+	l := NewCache(uint32(1000))
+	for i := 0; i < b.N; i++ {
+		l.Set(string(i), i)
+	}
+}
+
+func BenchmarkSetSize1Reuse(b *testing.B) {
+	l := NewCache(uint32(1))
+	l.Set("1", 1)
+	for i := 0; i < b.N; i++ {
+		l.Set("1", 1)
+	}
+}
+
+func BenchmarkGetSize1(b *testing.B) {
+	l := NewCache(uint32(1))
+	l.Set("1", 1)
+	for i := 0; i < b.N; i++ {
+		l.Get("1")
 	}
 }
